+++ app/store/wallet.ts
@@ -10,16 +10,11 @@ export interface WalletToken extends AddressToken {
 export interface WalletState {
   utxoBalance: string
   tokens: WalletToken[]
-  /**
-   * @deprecated use `useWalletContext().get(0).getAddress()` instead
-   */
-  address: string
 }
 
 const initialState: WalletState = {
   utxoBalance: '0',
-  tokens: [],
-  address: ''
+  tokens: []
 }
 
 const tokenDFI: WalletToken = {
@@ -61,9 +56,6 @@ export const wallet = createSlice({
     },
     setUtxoBalance: (state, action: PayloadAction<string>) => {
       state.utxoBalance = action.payload
-    },
-    setAddress: (state, action: PayloadAction<string>) => {
-      state.address = action.payload
     }
   }
 })
